# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wellcome to the Micopia Web Server Swagger Documentation",
        "title": "Micopia Swagger Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/CategoryCRUD/createCategory": {
            "put": {
                "description": "Create Category in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "description": "The Category to create",
                        "name": "CreateCategoryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    }
                }
            }
        },
        "/CategoryCRUD/deleteCategory": {
            "delete": {
                "description": "Delete Category in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "The Category identifier",
                        "name": "categoryId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/CategoryCRUD/getCategory": {
            "get": {
                "description": "Get Category from the database by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "The Category identifier",
                        "name": "categoryId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    }
                }
            }
        },
        "/CategoryCRUD/updateCategory": {
            "post": {
                "description": "Update Category in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "description": "The Category to update",
                        "name": "UpdateCategoryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Category": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CreateCategoryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}